Slescs/backend/seed_cert_and_tracking/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
crayons==0.4.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==3.2.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
grpcio==1.62.0
gunicorn==21.2.0
idna==3.6
IMAPClient==3.0.1
jsonrpclib==0.2.1
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
piexif==1.1.3
pillow==10.2.0
plumbum==1.8.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
rpyc==6.0.0
six==1.16.0
sqlparse==0.4.4
stegano==0.11.3
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
